+++ src/pushi/app/controllers/__init__.py
@@ -30,6 +30,7 @@ from . import app
 from . import base
 from . import event
 from . import health
+from . import messaging
 from . import smtp
 from . import socket
 from . import subscription
@@ -41,6 +42,7 @@ from .app import AppController
 from .base import BaseController
 from .event import EventController
 from .health import HealthController
+from .messaging import MessagingController
 from .smtp import SMTPController
 from .socket import SocketController
 from .subscription import SubscriptionController